+++ models/history/HIST_ABC_BANK_POSITION.sql
@@ -9,10 +9,9 @@
 {% set as_of_date_literal = "to_date('" ~ as_of_date ~ "')" %}
 {% set as_of_date_varchar_expr = "to_varchar(" ~ as_of_date_literal ~ ", 'YYYY-MM-DD')" %}
 
-
 {# Fields used to hash a synthetic CLOSE row (zeros + fixed date) #}
 {% set position_diff_fields_close = abc_bank_position_diff_fields(
-      prefix='curr.'
+      prefix='prior.'
     , report_date_expr=as_of_date_varchar_expr
     , quantity_expression='0'
     , cost_base_expression='0'
@@ -45,45 +44,52 @@ stg_input as (
 
 {% if is_incremental() %}
 
--- Current rows from target history table
-, current_from_history as (
-    {{ current_from_history(
-          history_rel = this
-        , key_column = 'POSITION_HKEY'
-        , history_filter_expr = "position_row_type = 'OPEN' and report_date < " ~ as_of_date_literal
-    ) }}
+-- Find the latest OPEN strictly before as_of_date, by business key
+, prior_latest_open as (
+  select *
+  from {{ this }}
+  where position_row_type = 'OPEN'
+    and report_date       < {{ as_of_date_literal }}
+  qualify row_number() over (
+           partition by position_hkey
+           order by
+                report_date desc
+              , load_ts_utc desc
+         ) = 1
+)
+
+, today_keys as (
+  select distinct position_hkey
+  from stg_today
 )
 
 -- Detect new rows (not already in history by diff hash)
 , new_rows as (
-    select stg.*
-    from stg_today as stg
-    left join current_from_history curr
-        on stg.position_hdiff = curr.position_hdiff
-    where curr.position_hdiff is null
+   select stg.*
+   from stg_today as stg
 )
 
 -- Detect positions missing from current batch and create synthetic CLOSE records
 , closed_positions as (
   select
-      curr.position_hkey                                                   as position_hkey
-    , {{ dbt_utils.generate_surrogate_key(position_diff_fields_close) }}   as position_hdiff
-    , curr.account_code                                                    as account_code
-    , curr.security_code                                                   as security_code
-    , curr.security_name                                                   as security_name
-    , curr.exchange_code                                                   as exchange_code
-    , curr.currency_code                                                   as currency_code
-    , curr.record_source                                                   as record_source
-    , {{ as_of_date_literal }}                                             as report_date
-    , 0                                                                    as quantity
-    , 0                                                                    as cost_base
-    , 0                                                                    as position_value
-    , '{{ run_started_at }}'::timestamp_ntz                                as load_ts_utc
-    , 'CLOSE_SYNTHETIC'                                                    as position_row_type
-  from current_from_history curr
-  left join stg_today stg
-    on stg.position_hkey = curr.position_hkey
-  where stg.position_hkey is null
+      prior.position_hkey                                                 as position_hkey
+    , {{ dbt_utils.generate_surrogate_key(position_diff_fields_close) }}  as position_hdiff
+    , prior.account_code                                                  as account_code
+    , prior.security_code                                                 as security_code
+    , prior.security_name                                                 as security_name
+    , prior.exchange_code                                                 as exchange_code
+    , prior.currency_code                                                 as currency_code
+    , prior.record_source                                                 as record_source
+    , {{ as_of_date_literal }}                                            as report_date
+    , 0                                                                   as quantity
+    , 0                                                                   as cost_base
+    , 0                                                                   as position_value
+    , '{{ run_started_at }}'::timestamp_ntz                               as load_ts_utc
+    , 'CLOSE_SYNTHETIC'                                                   as position_row_type
+  from prior_latest_open prior
+  left join today_keys tk
+    on tk.position_hkey = prior.position_hkey
+  where tk.position_hkey is null
 )
 
 , new_rows_projected as (